requirements.txt
google-genai==1.7.0
requests==2.31.0
